+++ packages/preview-server/src/contexts/templates.tsx
@@ -1,10 +1,10 @@
 'use client';
 
+import type { TemplatesDirectory } from '@ahmedrowaihi/pdf-forge-templates';
 import { createContext, useContext, useState } from 'react';
 import { getTemplatesDirectoryMetadataAction } from '../actions/get-templates-directory-metadata-action';
 import { isBuilding, isPreviewDevelopment } from '../app/env';
 import { useHotreload } from '../hooks/use-hot-reload';
-import type { TemplatesDirectory } from '../utils/get-templates-directory-metadata';
 
 const TemplatesContext = createContext<
   | {