+++ packages/templates/src/index.ts
@@ -1,0 +1,2 @@
+export type { TemplatesDirectory } from './get-templates-directory-metadata.js';
+export { getTemplatesDirectoryMetadata } from './get-templates-directory-metadata.js';